Patching allocate function - 05Sep2025 (#28)
* Patching allocate function - 05Sep2025

This includes a few tweaks. First, we generate a warning if sum(N.h) < n.samp. Next, I tweaked the formula for the denominator of the optimal allocation (cost-constrained). I also made sure that n is at least length(N.h)*lbound (mostly because of issues with the optimal case). And finally, I modified how the remaining sizes are allocated, to fix an issue where rounding would yield the wrong final allocations.

* Making a few tweaks to allocate() and adding testing for allocate() - 14Sep2025

I made a few more tweaks to the allocate() function: I added a few all() calls to checks of the inputs - this will help if the user provides more than one value when expected (e.g., if they put cost as a vector of length >1). For similar reasons, I included more explicit checks of the inputs when doing the "lbound*length(N.h) must be less than or equal to n.samp" check - I added an if() regarding all 3 of lbound, N.h, and n.samp not being null before proceeding (I also added an all() to the check that comes after). Furthermore, I added similar explicit checks for the "sum(N.h) is less than n.samp" check. And I added a new check for the optimal (cost) method - I throw a warning if no solution exists at/below the 'cost' parameter (e.g., sum(lbound*c.h) > cost). The last final change is that I tweaked the denominator of the optimal (variance) formula to be more in line with the specs. Other than that, I tweaked indenting in a few places.

The bigger change in this update is the addition of testing for the allocate() function. These match my proposed testing document (which I also updated today). This includes multiple tests, each with multiple cases, for each of the allocation methods. There's still room for more testing, but I think this is a very solid starting point.